添加 请求参数校验

+++ wish-xinmi-nodets-server/src/router/contact-router.ts
@@ -3,18 +3,32 @@ import R from "../util/response";
 import userService from "../service/user-service";
 import socket from "../socket";
 import {routerFactory} from "./router-factory";
+import Joi from 'joi';
 
 const router = routerFactory("/contact");
 
 router.get("/getYesContactList", async (ctx: any) => {
     const query = ctx.query;
-    const userId = query.userId;
+    const {userId} = query;
+
+    if (!userService.hasUser(userId)) {
+        throw Error('找不到用户');
+    }
+
     const list = await contactService.getYesContactList(userId);
     ctx.body = R.success(list);
 });
 
 router.get("/addContact", async (ctx: any) => {
     const {userId, contactId, validateMsg} = ctx.query;
+
+    if (!userService.hasUser(userId)) {
+        throw Error(`找不到${userId}`);
+    }
+
+    if (!userService.hasUser(contactId)) {
+        throw Error(`找不到${contactId}`);
+    }
 
     const status: any = await contactService.addContact(userId, contactId, validateMsg);
 
@@ -27,8 +41,12 @@ router.get("/addContact", async (ctx: any) => {
 
 router.get("/getNoContactList", async (ctx: any) => {
     const query = ctx.query;
-    const userId = query.userId;
-    const username = query.username || "";
+    const {userId, username = ''} = query;
+
+    if (!userService.hasUser(userId)) {
+        throw Error('找不到用户');
+    }
+
     const list = await contactService.getNoContactList(userId, username);
     ctx.body = R.success(list);
 });
@@ -36,30 +54,57 @@ router.get("/getNoContactList", async (ctx: any) => {
 router.get("/getConfirmContactList", async (ctx: any) => {
     const query = ctx.query;
     const userId = query.userId;
+
+    if (!userService.hasUser(userId)) {
+        throw Error('找不到用户');
+    }
+
     const list = await contactService.getConfirmContactList(userId);
     ctx.body = R.success(list);
 });
 
 router.get("/confirmContact", async (ctx: any) => {
     const query = ctx.query;
-    const userId = query.userId;
-    const contactId = query.contactId;
+    const {userId, contactId} = query;
+
+    if (!userService.hasUser(userId)) {
+        throw Error('找不到用户');
+    }
+
+    if (!userService.hasUser(userId)) {
+        throw Error('找不到用户');
+    }
+
     await contactService.confirmContact(userId, contactId);
     ctx.body = R.success();
 });
 
 router.get("/getUserContactStatus", async (ctx: any) => {
     const query = ctx.query;
-    const userId = query.userId;
-    const contactId = query.contactId;
+    const {userId, contactId} = query;
+    if (!userService.hasUser(userId)) {
+        throw Error('找不到用户' + userId);
+    }
+
+    if (!userService.hasUser(contactId)) {
+        throw Error('找不到用户' + contactId);
+    }
+
     const status = await contactService.getUserContactStatus(userId, contactId);
     ctx.body = R.success(status);
 });
 
 router.get("/getContactInfoHad", async (ctx: any) => {
     const query = ctx.query;
-    const userId = query.userId;
-    const contactId = query.contactId;
+    const {userId, contactId} = query;
+    if (!userService.hasUser(userId)) {
+        throw Error('找不到用户' + userId);
+    }
+
+    if (!userService.hasUser(contactId)) {
+        throw Error('找不到用户' + contactId);
+    }
+
     const info = await contactService.getContactInfoHad(userId, contactId);
     ctx.body = R.success(info);
 });
@@ -69,6 +114,11 @@ router.get("/getContactInfoHad", async (ctx: any) => {
  */
 router.get("/getContactWarnNum", async (ctx: any) => {
     const {userId} = ctx.query;
+
+    if (!userService.hasUser(userId)) {
+        throw Error('找不到用户' + userId);
+    }
+
     const num = await contactService.getContactWarnNum({userId});
     ctx.body = R.success(num);
 });
@@ -78,6 +128,11 @@ router.get("/getContactWarnNum", async (ctx: any) => {
  */
 router.post("/setAllContactChecked", async (ctx: any) => {
     const {userId} = ctx.request.body;
+
+    if (!userService.hasUser(userId)) {
+        throw Error('找不到用户' + userId);
+    }
+
     await contactService.setAllContactChecked(userId);
     ctx.body = R.success();
 });
@@ -87,6 +142,16 @@ router.post("/setAllContactChecked", async (ctx: any) => {
  */
 router.post("/editContact", async (ctx: any) => {
     const {userId, contactId, contactName} = ctx.request.body;
+
+    if (!userService.hasUser(userId)) {
+        throw Error('找不到用户' + userId);
+    }
+
+    if (!userService.hasUser(contactId)) {
+        throw Error('找不到用户' + contactId);
+    }
+
+
     await contactService.editContact(userId, contactId, contactName);
     ctx.body = R.success();
 });
@@ -96,6 +161,18 @@ router.post("/editContact", async (ctx: any) => {
  */
 router.post("/deleteContact", async (ctx: any) => {
     const {userId, contactId} = ctx.request.body;
+    if (!userService.hasUser(userId)) {
+        throw Error('找不到用户' + userId);
+    }
+
+    if (!userService.hasUser(contactId)) {
+        throw Error('找不到用户' + contactId);
+    }
+
+    if (!contactService.isContact(userId, contactId)) {
+        throw Error(`${userId}没有联系人${contactId}`);
+    }
+
     await contactService.deleteContact(userId, contactId);
     ctx.body = R.success();
 });
